+++ resources/views/components/emails/footer-html.blade.php
@@ -1,0 +1,12 @@
+<hr />
+@php
+    $providerLink = sprintf(
+        '<a href="https://%s">%s</a>',
+        __('layouts.provider.www'),
+        __('layouts.provider.name')
+    );
+@endphp
+<p>
+    <x-ig-common::emails.service />
+    @lang('layouts.provider', ['link' => $providerLink, 'year' => date('Y')])
+</p>