+++ src/CommonServiceProvider.php
@@ -12,6 +12,10 @@ use InternetGuru\LaravelCommon\Exceptions\Handler;
 use InternetGuru\LaravelCommon\Livewire\Messages;
 use InternetGuru\LaravelCommon\Services\ReCaptcha;
 use Livewire\Livewire;
+use Illuminate\Support\Facades\Event;
+use Illuminate\Notifications\Events\NotificationSent;
+use InternetGuru\LaravelCommon\Listeners\LogSentNotification;
+
 
 class CommonServiceProvider extends ServiceProvider
 {
@@ -35,6 +39,15 @@ class CommonServiceProvider extends ServiceProvider
         $this->loadTranslationsFrom(__DIR__ . '/../lang', 'ig-common');
         Blade::componentNamespace('InternetGuru\LaravelCommon\View\Components', 'ig');
         Livewire::component('ig-messages', Messages::class);
+
+        $this->publishes([
+            __DIR__ . '/../database/migrations' => database_path('migrations'),
+        ], 'ig-common:migrations');
+
+        Event::listen(
+            NotificationSent::class,
+            [LogSentNotification::class, 'handle']
+        );
 
         // Register middleware to inject ReCaptcha script
         $this->app['router']->pushMiddlewareToGroup('web', InjectRecaptchaScript::class);