+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2025-05-15
+
+_Stable release based on [2.0.0-rc.1]._
+
+## [2.0.0-rc.1] - 2025-05-15
+
+### Added
+
+- Add `MailMessage` that add ref to subject.
+- Log emails into database and to standard log using notification sent listener.
+
 ## [1.0.4] - 2025-05-09
 
 ### Fixed
@@ -449,6 +460,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[2.0.0]: https://https://github.com/internetguru/laravel-common/compare/v1.0.4...v2.0.0
+[2.0.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v1.0.4
 [1.0.4]: https://https://github.com/internetguru/laravel-common/compare/v1.0.3...v1.0.4
 [1.0.3]: https://https://github.com/internetguru/laravel-common/compare/v1.0.2...v1.0.3
 [1.0.2]: https://https://github.com/internetguru/laravel-common/compare/v1.0.1...v1.0.2